Move components to app
