+++ app/artifacts/components/ArtifactData.tsx
