+++ app/artifacts/components/ArtifactIndex.tsx
