+++ app/artifacts/components/WindowHeader.tsx
