+++ app/components/flux/CubeFiber.tsx
